Pokemon_API (Database)/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.1
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
dnspython==2.7.0
docutils==0.21.2
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.5
filelock==3.0.12
greenlet==3.1.1
h11==0.14.0
idna==2.8
jose==1.0.0
jwt==1.3.1
lockfile==0.12.2
Mako==1.3.8
MarkupSafe==3.0.2
multipart==1.2.1
passlib==1.7.4
psycopg2==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-daemon==2.2.4
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
requests==2.22.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
urllib3==1.25.11
uvicorn==0.32.0